Preview binarized results. Implement size variation in image viewing.
Implement image previewing generally. Specifically, allow for previewing of the binarization results. Outfit a pipeline stage to update previews, since loading the preview from disk each time is slow. Outside of the pipeline, set up a task to run the binarized preview.

To enable the above, allow image size variation in a single image view in the GUI.