Refactor out model functionality pertaining to model synchronization
Create `SynchronizedPathsYamlModel`, a model of a YAML file that synchronizes paths back to the file. Create `DefaultPathsModel`, a model in which all fields must be path-like and have defaults specified in this model. Create a derived `CreatePathsModel` which inherits from `DefaultPathsModel` and additionally, creates parent directories for all fields in this model.

Retire the `schema` and synchronization stages. Sync is now automatic whenever the `params.yaml` model is created at runtime. This means that `Params` now reflects the state of `params.yaml` at loading, by ingesting specified parameters, and importantly, back-propagating constant paths back to it. This facilitates passing relevant project paths back to DVC, the pipeline orchestrator. This two-way sync is necessary because path composition by YAML alone is unfriendly and static, e.g. paths are not automatically re-created, and code generation would be necessary to sync in the forward direction.